+++ src/010-pipeline/build.ts
@@ -9,8 +9,8 @@ import { writeReport } from "../080-generator/writeReport";
 import { getRawAST } from "../050-grammar/store";
 import files from "../040-imports/stdlib";
 import { ILogger, Logger } from "./logger";
-import { PackageFileFormat } from "../packaging/fileFormat";
-import { packageCode } from "../packaging/packageCode";
+import { PackageFileFormat } from "../110-packaging/fileFormat";
+import { packageCode } from "../110-packaging/packageCode";
 import { createABITypeRefFromTypeRef } from "../060-types/resolveABITypeRef";
 import { getContracts, getType } from "../060-types/resolveDescriptors";
 import { posixNormalize } from "../utils/filePath";
